LayoutTests/platform/chromium/fast/events/onunload-clears-onbeforeunload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: Blocked alert('before unload') during beforeunload.
CONSOLE MESSAGE: Blocked alert('unload') during unload.
If you didn't hit an assert you PASS.
